KDTF_EB/requirements.txt
selenium
webdriver-manager
pytest
pytest-html
pyyaml